fix(grpc-sdk): health check middleware injection fix(grpc-sdk): middlewares not injected in module clients refactor(core): error logging in module recovery
