+++ libraries/grpc-sdk/src/classes/HealthCheck.ts
@@ -1,6 +1,6 @@
 import { getGrpcSignedTokenInterceptor, getModuleNameInterceptor } from '../interceptors';
 import { createChannel, createClientFactory } from 'nice-grpc';
-import { HealthCheckResponse, HealthDefinition } from '../protoUtils/grpc_health_check';
+import { HealthCheckResponse, HealthDefinition } from '../protoUtils';
 import ConduitGrpcSdk from '../index';
 
 export async function checkModuleHealth(
@@ -13,15 +13,14 @@ export async function checkModuleHealth(
     'grpc.max_receive_message_length': 1024 * 1024 * 100,
     'grpc.max_send_message_length': 1024 * 1024 * 100,
   });
-  const clientFactory = createClientFactory();
-  for (const interceptor of ConduitGrpcSdk.interceptors) {
-    clientFactory.use(interceptor);
-  }
-  clientFactory.use(
+  let clientFactory = createClientFactory().use(
     grpcToken
       ? getGrpcSignedTokenInterceptor(grpcToken)
       : getModuleNameInterceptor(clientName),
   );
+  for (const interceptor of ConduitGrpcSdk.interceptors) {
+    clientFactory = clientFactory.use(interceptor);
+  }
   const _healthClient = clientFactory.create(HealthDefinition, channel);
 
   let error;
